requirements.txt
alpaca-trade-api==3.0.0
python-dotenv==1.0.0
apscheduler==3.10.1
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
yfinance==0.2.65
pytz==2023.3
matplotlib==3.7.2
seaborn==0.12.2 